+++ app/src/main/java/eu/kanade/tachiyomi/ui/setting/controllers/debug/DebugController.kt
@@ -4,15 +4,15 @@ import android.os.Build
 import androidx.preference.PreferenceScreen
 import androidx.webkit.WebViewCompat
 import eu.kanade.tachiyomi.BuildConfig
-import eu.kanade.tachiyomi.ui.more.AboutController
 import eu.kanade.tachiyomi.ui.setting.SettingsLegacyController
 import eu.kanade.tachiyomi.ui.setting.onClick
 import eu.kanade.tachiyomi.ui.setting.preference
 import eu.kanade.tachiyomi.ui.setting.preferenceCategory
 import eu.kanade.tachiyomi.util.system.DeviceUtil
 import eu.kanade.tachiyomi.util.view.withFadeTransaction
-import yokai.i18n.MR
 import java.text.DateFormat
+import yokai.i18n.MR
+import yokai.presentation.settings.screen.about.getFormattedBuildTime
 
 class DebugController : SettingsLegacyController() {
 
@@ -49,7 +49,7 @@ class DebugController : SettingsLegacyController() {
             preference {
                 key = "pref_build_time"
                 title = "Build Time"
-                summary = AboutController.getFormattedBuildTime(dateFormat)
+                summary = getFormattedBuildTime(dateFormat)
             }
             preference {
                 key = "pref_webview_version"