+++ presentation/core/src/main/java/yokai/presentation/core/icons/CustomIcons.kt
@@ -1,0 +1,3 @@
+package yokai.presentation.core.icons
+
+object CustomIcons